Prevent crash with bugged inventories
Never came across this myself, but it was reported in Steam comments and adding a nil check here is harmless.

This means that now instead of crashing when announcing bugged items with perishability, it will just fail to announce how fresh they are. Probably. Unless the buggedness causes it to crash on something else later :P