Improve the per-announcement cooldown system
Before, it was keyed off the of the exact message. This meant that for things that are changing quickly, you can really spam announcements of them (for example cooldowns on Wanda's clocks), but only if you had the option on to show exact numbers.

This changes it to be keyed off of the type of message (a particular stat, item, etc), which should be consistent regardless of the message content.