Add "Hide Announcements" option
I've been told that some people are bothered by the ability of this mod to affect their experience on random servers (because it's a client mod that can be used anywhere). So this should do a lot to help by preventing them from seeing any announcements. Behavior changes of other players, though... nothing to be done about that.